hawao: Adjust horizontal position of statusbar cutout
The cutout_space_view is shifted left by 20 pixel. That wastes space
on the left side of the cutout, and the right content part extends
into the cutout area.
An additional padding of 12dp on the left side shifts everything into
position.

And the paddings on left and right are slightly increased, to fit also
the screen module edovtdr_ft3519_643_video, which has corner circles
with larger radius than the screen module visionoxrm_ft3519_643_video
which is used alternatively.

Change-Id: Ia876d4e36b519d8eba41fc370ed4c0048bdd0d1f